refactor: controller owns its API contract — shared/contracts becomes controller/contracts
The @local-studio/contracts package moves inside the controller (the
server owns the contract; clients consume it). The package name and
every import statement stay unchanged: the frontend's file: dependency
repoints to controller/contracts, the tsconfig aliases follow, and the
two relative imports (shared/agent, agent-runtime) are repointed. The
contracts validator scans the new home.
