+++ controller/contracts/package.json
@@ -4,6 +4,6 @@
   "private": true,
   "type": "module",
   "exports": {
-    "./*": "./contracts/*.ts"
+    "./*": "./*.ts"
   }
 }
