+++ controller/contracts/speech.ts
