+++ controller/tests/integration/log-redaction.test.ts
