+++ controller/tests/integration/model-catalog-contracts.test.ts
