+++ controller/tests/integration/openai-routes-helpers.test.ts
