+++ controller/tests/integration/tool-call-stream.test.ts
