+++ controller/tests/integration/vllm-docker-image.test.ts
