+++ frontend/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "predev": "node scripts/link-services-node-modules.mjs",
     "dev": "next dev",
-    "build": "next build --webpack",
+    "build": "next build",
     "start": "node scripts/start-standalone.mjs",
     "perf:audit": "node scripts/perf-audit.mjs",
     "perf:browser": "node scripts/browser-perf-audit.mjs",
@@ -44,7 +44,7 @@
     "@earendil-works/pi-coding-agent": "0.78.1",
     "@hono/node-server": "1.19.14",
     "@local-studio/agent-runtime": "file:../services/agent-runtime",
-    "@local-studio/contracts": "file:../shared",
+    "@local-studio/contracts": "file:../controller/contracts",
     "@lydell/node-pty": "1.2.0-beta.12",
     "@xterm/addon-fit": "0.11.0",
     "@xterm/addon-web-links": "0.13.0-beta.220",
